project/requirements.txt
Django==4.2
python-dotenv==1.0.0
django-split-settings==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
gunicorn==20.1.0
django-crispy-forms==1.14.0
psycopg2-binary==2.9.6
channels==3.0.5
pydantic==1.9.1
motor==3.1.1
uvicorn[standard]==0.19.0
